+++ ql/src/java/org/apache/hadoop/hive/ql/optimizer/calcite/rules/HivePointLookupOptimizerRule.java
@@ -22,10 +22,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.IdentityHashMap;
-import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
@@ -52,6 +49,7 @@ import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.hadoop.hive.ql.optimizer.calcite.HiveCalciteUtil;
 import org.apache.hadoop.hive.ql.optimizer.calcite.reloperators.HiveBetween;
 import org.apache.hadoop.hive.ql.optimizer.calcite.reloperators.HiveIn;
+import org.apache.hadoop.hive.ql.optimizer.graph.DiGraph;
 import org.apache.hadoop.hive.ql.parse.SemanticException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -188,167 +186,6 @@ public abstract class HivePointLookupOptimizerRule extends RelOptRule {
     RexTransformIntoBetween t = new RexTransformIntoBetween(rexBuilder);
     newCondition = t.apply(newCondition);
     return newCondition;
-  }
-
-  public static class DiGraph<V, E> {
-    protected static class Edge<V, E> {
-      public final Node<V, E> s;
-      public final Node<V, E> t;
-      public final E e;
-
-      public Edge(Node<V, E> s, Node<V, E> t, E e) {
-        this.s = s;
-        this.t = t;
-        this.e = e;
-      }
-    }
-
-    protected static class Node<V, E> {
-      public final Set<Edge<V, E>> edges;
-      public final V v;
-
-      public Node(V v) {
-        edges = new LinkedHashSet<>();
-        this.v = v;
-      }
-
-      public void addEdge(Edge<V, E> edge) {
-        if (edge.s == this) {
-          edge.t.addEdge(edge);
-        }
-        edges.add(edge);
-      }
-
-      public E removeEdge(V s, V t) {
-        Iterator<Edge<V, E>> it = edges.iterator();
-        while (it.hasNext()) {
-          Edge<V, E> edge = it.next();
-          if (edge.s.v.equals(s) && edge.t.v.equals(t)) {
-            it.remove();
-            return edge.e;
-          }
-        }
-        return null;
-      }
-
-      public int degree() {
-        return edges.size();
-      }
-
-      public int inDegree() {
-        Iterator<Edge<V, E>> it = edges.iterator();
-        int cnt = 0;
-        while (it.hasNext()) {
-          Edge<V, E> edge = it.next();
-          if (edge.t == this) {
-            cnt++;
-          }
-        }
-        return cnt;
-      }
-
-
-
-    }
-
-    private final Map<V, Node<V, E>> nodes;
-
-    public DiGraph() {
-      nodes = new LinkedHashMap<>();
-    }
-
-    public void putEdgeValue(V s, V t, E e) {
-      Node<V, E> nodeS = nodeOf(s);
-      Node<V, E> nodeT = nodeOf(t);
-      Edge<V, E> edge = new Edge<>(nodeS, nodeT, e);
-      nodeS.addEdge(edge);
-      //      nodeT.addEdge(edge);
-    }
-
-    public Node<V, E> nodeOf(V s) {
-      Node<V, E> node = nodes.get(s);
-      if (node == null) {
-        nodes.put(s, node = newNode(s));
-      }
-      return node;
-    }
-
-    protected Node<V, E> newNode(V s) {
-      return new Node<>(s);
-    }
-
-    public Set<V> nodes() {
-      return nodes.keySet();
-    }
-
-    public Set<V> predecessors(V n) {
-      Set<V> ret = new LinkedHashSet<>();
-      Node<V, E> node = nodes.get(n);
-      if (node == null) {
-        return ret;
-      }
-
-      for (Edge<V, E> edge : node.edges) {
-        if (edge.t.v.equals(n)) {
-          ret.add(edge.s.v);
-        }
-      }
-      return ret;
-    }
-
-    public Set<V> successors(V n) {
-      Set<V> ret = new LinkedHashSet<>();
-      Node<V, E> node = nodes.get(n);
-      if (node == null) {
-        return ret;
-      }
-
-      for (Edge<V, E> edge : node.edges) {
-        if (edge.s.v.equals(n)) {
-          ret.add(edge.t.v);
-        }
-      }
-      return ret;
-    }
-
-    public int degree(V n) {
-      Node<V, E> node = nodes.get(n);
-      return node.degree();
-    }
-
-    public int inDegree(V n) {
-      Node<V, E> node = nodes.get(n);
-      return node.inDegree();
-    }
-
-    public E removeEdge(V s, V t) {
-      nodeOf(s).removeEdge(s, t);
-      return nodeOf(t).removeEdge(s, t);
-    }
-
-    public void impode(V n, E edge) {
-      Set<V> preds = predecessors(n);
-      Set<V> succ = successors(n);
-
-      for (V u : preds) {
-        for (V v : succ) {
-          putEdgeValue(u, v, edge);
-        }
-      }
-      delete(n);
-
-
-    }
-
-    public void delete(V n) {
-      Node<V, E> node = nodeOf(n);
-      Set<Edge<V, E>> edges = new HashSet<>(node.edges);
-      for (Edge<V, E> edge : edges) {
-        removeEdge(edge.s.v, edge.t.v);
-      }
-      nodes.remove(n);
-    }
-
   }
 
   /**
